[Fix] Animations
-Animatable class removed.
-Animations applied directly to objects.
