+++ scripts/planet.js
@@ -1,21 +1,76 @@
 import { container } from './GameManager.js';
-import { Animatable } from './Animatable.js';
 
-export class Planet extends Animatable
+export class Planet
 {
     constructor()
     {
-        super();
+        this.animatedSprites = {};
+        this.currentAnimatedSprite;
         this.animationSpeed = 0.5;
-        this.spritePath = 'sprites/moon.json'
+        this.spritePath = 'sprites/moon.json';
         this.spriteWidth = 100;
         this.spriteHeight = 100;
         this.loadSpriteSheet();
     }
 
+    //Carga el spritesheet del objeto basado en el "spritePath".
     async loadSpriteSheet()
     {
-        await super.loadSpriteSheet();
-        //console.log(this.currentAnimatedSprite);
+        await PIXI.Assets.load(this.spritePath).then(sheet => this.initAnimationsLoad(sheet));
+    }
+
+    //Crea un set de animaciones en base a lo descripto en el "spritePath" e iniciliza la animacion "mooving".
+    initAnimationsLoad(sheet)
+    {
+        this.animatedSprites = 
+        {
+            'none' : new PIXI.AnimatedSprite(sheet.animations['none']),
+            'mooving' : new PIXI.AnimatedSprite(sheet.animations['mooving']),
+            'beHarmed' : new PIXI.AnimatedSprite(sheet.animations['beHarmed']),
+            'vfx' : new PIXI.AnimatedSprite(sheet.animations['vfx']),
+        }
+
+        const animatedSpritesValue = Object.values(this.animatedSprites);
+        animatedSpritesValue.forEach(s => this.prepareToBeUsed(s));
+
+        this.play('mooving');
+    }
+
+    //Setea los valores correpondientes al tamaño del sprite.
+    prepareToBeUsed(animatedSprite)
+    {
+        animatedSprite.width = this.spriteWidth;
+        animatedSprite.height = this.spriteHeight;
+        animatedSprite.anchor = (0.5, 0.5);
+        animatedSprite.animationSpeed = this.animationSpeed;
+    }
+
+    //Ejecuta una animación perteneciente a dicionario "animatedSprites".
+    play(animationKey)
+    {
+        this.currentAnimatedSprite = this.animatedSprites[animationKey];
+        container.addChild(this.currentAnimatedSprite);
+        this.currentAnimatedSprite.play();
+    }
+
+    //Para la animación que se está ejecutando actualmente.
+    stop()
+    {
+        this.currentAnimatedSprite.stop();
+    }
+
+    //Cambia la animación actual por la que recibe por parámetro.
+    changeAnimation(animationKey)
+    {
+        container.removeChild(this.currentAnimatedSprite);
+        this.currentAnimatedSprite = this.animatedSprites[animationKey];
+        container.addChild(this.currentAnimatedSprite);
+        this.currentAnimatedSprite.play();
+    }
+
+    //Cambia la velocidad de la animación.
+    changeAnimationSpeed(speed)
+    {
+        this.currentAnimatedSprite.animationSpeed = speed;
     }
 }
